Added: -AsynchronousTask Factory -FetchReducedWeatherTask -FetchExtendedWeatherTask
