+++ app/src/main/java/com/example/cicerone/first/Fragment/ForecastFragment.java
@@ -3,12 +3,9 @@ package com.example.cicerone.first.Fragment;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.net.Uri;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -16,29 +13,14 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
 import com.example.cicerone.first.Activity.DetailActivity;
 import com.example.cicerone.first.Adapter.ItemAdapter;
 import com.example.cicerone.first.Domain.Weather;
+import com.example.cicerone.first.Factory.FetchWeatherTaskFactory;
+import com.example.cicerone.first.Helper.AFetchWeatherTask;
 import com.example.cicerone.first.R;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.Console;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * Encapsulates fetching the forecast and displaying it as a {@link ListView} layout.
@@ -46,7 +28,7 @@ import java.util.List;
 public class ForecastFragment extends Fragment {
 
     private ItemAdapter mForecastAdapter;
-    private ArrayList<Weather> weekForecast;
+
 
     public ForecastFragment() {
     }
@@ -63,7 +45,7 @@ public class ForecastFragment extends Fragment {
     }
 
     public void goWeather(){
-        FetchWeatherTask fwt = new FetchWeatherTask();
+        AFetchWeatherTask fwt = (new FetchWeatherTaskFactory()).makeFetchWeatherTask(false,mForecastAdapter);
         SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
         String lat = prefs.getString(getString(R.string.pref_latitude_key),getString(R.string.pref_latitude_default));
         String lon = prefs.getString(getString(R.string.pref_longitude_key),getString(R.string.pref_longitude_default));
@@ -74,7 +56,7 @@ public class ForecastFragment extends Fragment {
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
         if(id == R.id.action_refresh){
-            goWeather();
+            this.goWeather();
             return true;
         }
         return super.onOptionsItemSelected(item);
@@ -83,7 +65,7 @@ public class ForecastFragment extends Fragment {
     @Override
     public void onStart() {
         super.onStart();
-        goWeather();
+
     }
 
     @Override
@@ -103,6 +85,9 @@ public class ForecastFragment extends Fragment {
         // Get a reference to the ListView, and attach this adapter to it.
         ListView listView = (ListView) rootView.findViewById(R.id.listview_forecast);
         listView.setAdapter(mForecastAdapter);
+
+        this.goWeather();
+
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
@@ -119,201 +104,4 @@ public class ForecastFragment extends Fragment {
         return rootView;
     }
 
-    public class FetchWeatherTask extends AsyncTask<String, Void, Weather[]> {
-
-        private final String LOG_TAG = FetchWeatherTask.class.getSimpleName();
-
-        @Override
-        protected void onPostExecute(Weather[] result) {
-            if (result!=null){
-                mForecastAdapter.clear();
-                for (Weather singleDay : result){
-                    mForecastAdapter.addItem(singleDay);
-                }
-                //Ehi! I'm changed
-                mForecastAdapter.notifyDataSetChanged();
-            }
-        }
-
-        /* The date/time conversion code is going to be moved outside the asynctask later,
-         * so for convenience we're breaking it out into its own method now.
-         */
-        private String getReadableDateString(long time){
-            // Because the API returns a unix timestamp (measured in seconds),
-            // it must be converted to milliseconds in order to be converted to valid date.
-            java.util.Date date = new java.util.Date(time * 1000);
-            SimpleDateFormat format = new SimpleDateFormat("E d MMM");
-            return format.format(date).toString();
-        }
-
-        /**
-         * Prepare the weather high/lows for presentation.
-         */
-        private String formatHighLows(double high, double low) {
-            //check if the unit is imperial
-            SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
-            String unit = prefs.getString(getString(R.string.pref_unit_key),getString(R.string.pref_unit_default));
-            if (unit.equals(getString(R.string.pref_units_imperial))){
-                high = (high * 1.8) + 32;
-                low = (low * 1.8) +32;
-            }else if(!unit.equals(getString(R.string.pref_units_metric))){
-                Log.e(LOG_TAG,"Unit not found");
-            }
-
-            // For presentation, assume the user doesn't care about tenths of a degree.
-            long roundedHigh = Math.round(high);
-            long roundedLow = Math.round(low);
-            String highLowStr = roundedHigh + "/" + roundedLow;
-            return highLowStr;
-        }
-
-        /**
-         * Take the String representing the complete forecast in JSON Format and
-         * pull out the data we need to construct the Strings needed for the wireframes.
-         *
-         * Fortunately parsing is easy:  constructor takes the JSON string and converts it
-         * into an Object hierarchy for us.
-         */
-        private Weather[] getWeatherDataFromJson(String forecastJsonStr, int numDays)
-                throws JSONException {
-
-            // These are the names of the JSON objects that need to be extracted.
-            final String OWM_LIST = "list";
-            final String OWM_WEATHER = "weather";
-            final String OWM_TEMPERATURE = "temp";
-            final String OWM_MAX = "max";
-            final String OWM_MIN = "min";
-            final String OWM_DATETIME = "dt";
-            final String OWM_DESCRIPTION = "main";
-            final String OWM_ICON = "icon";
-
-            JSONObject forecastJson = new JSONObject(forecastJsonStr);
-            JSONArray weatherArray = forecastJson.getJSONArray(OWM_LIST);
-
-            Weather[] result_data = new Weather[numDays];
-            Weather weather;
-            for(int i = 0; i < weatherArray.length(); i++) {
-                // For now, using the format "Day, description, hi/low"
-                weather = new Weather();
-
-                // Get the JSON object representing the day
-                JSONObject dayForecast = weatherArray.getJSONObject(i);
-
-                // The date/time is returned as a long.  We need to convert that
-                // into something human-readable, since most people won't read "1400356800" as
-                // "this saturday".
-                long dateTime = dayForecast.getLong(OWM_DATETIME);
-                weather.setDatetime(getReadableDateString(dateTime));
-
-                // description is in a child array called "weather", which is 1 element long.
-                JSONObject weatherObject = dayForecast.getJSONArray(OWM_WEATHER).getJSONObject(0);
-                weather.setDescription(weatherObject.getString(OWM_DESCRIPTION));
-
-                JSONObject IconObject = dayForecast.getJSONArray(OWM_WEATHER).getJSONObject(0);
-                weather.setImage(IconObject.getString(OWM_ICON));
-
-                // Temperatures are in a child object called "temp".  Try not to name variables
-                // "temp" when working with temperature.  It confuses everybody.
-                JSONObject temperatureObject = dayForecast.getJSONObject(OWM_TEMPERATURE);
-                weather.setMax(String.valueOf(temperatureObject.getDouble(OWM_MAX)));
-                weather.setMin(String.valueOf(temperatureObject.getDouble(OWM_MIN)));
-
-                //highAndLow = formatHighLows(high, low);
-                result_data[i] = weather;
-
-            }
-
-
-
-            return result_data;
-        }
-
-        @Override
-        protected Weather[] doInBackground(String... params) {
-            // These two need to be declared outside the try/catch
-            // so that they can be closed in the finally block.
-            HttpURLConnection urlConnection = null;
-            BufferedReader reader = null;
-
-            // Will contain the raw JSON response as a string.
-            String forecastJsonStr = null;
-
-            String format = "json";
-            String units = "metric";
-            int numDays = 7;
-            for (String param : params)
-            {Log.v(LOG_TAG, param);}
-
-
-            try {
-
-                final String FORECAST_BASE_URL = "http://api.openweathermap.org/data/2.5/forecast/daily?";
-                final String LAT_PARAM = "lat";
-                final String LON_PARAM = "lon";
-                final String FORMAT_PARAM = "mode";
-                final String UNITS_PARAM = "units";
-                final String DAYS_PARAM = "cnt";
-
-                Uri builtUri = Uri.parse(FORECAST_BASE_URL).buildUpon()
-                        .appendQueryParameter(LAT_PARAM,params[0])
-                        .appendQueryParameter(LON_PARAM,params[1])
-                        .appendQueryParameter(FORMAT_PARAM,format)
-                        .appendQueryParameter(UNITS_PARAM,units)
-                        .appendQueryParameter(DAYS_PARAM,Integer.toString(numDays))
-                        .build();
-
-                URL url = new URL(builtUri.toString());
-
-                // Create the request to OpenWeatherMap, and open the connection
-                urlConnection = (HttpURLConnection) url.openConnection();
-                urlConnection.setRequestMethod("GET");
-                urlConnection.connect();
-
-                // Read the input stream into a String
-                InputStream inputStream = urlConnection.getInputStream();
-                StringBuffer buffer = new StringBuffer();
-                if (inputStream == null) {
-                    // Nothing to do.
-                    return null;
-                }
-                reader = new BufferedReader(new InputStreamReader(inputStream));
-
-                String line;
-                while ((line = reader.readLine()) != null) {
-                    // Since it's JSON, adding a newline isn't necessary (it won't affect parsing)
-                    // But it does make debugging a *lot* easier if you print out the completed
-                    // buffer for debugging.
-                    buffer.append(line + "\n");
-                }
-
-                if (buffer.length() == 0) {
-                    // Stream was empty.  No point in parsing.
-                    return null;
-                }
-                forecastJsonStr = buffer.toString();
-            } catch (IOException e) {
-                Log.e(LOG_TAG, "Error ", e);
-                // If the code didn't successfully get the weather data, there's no point in attemping
-                // to parse it.
-                return null;
-            } finally {
-                if (urlConnection != null) {
-                    urlConnection.disconnect();
-                }
-                if (reader != null) {
-                    try {
-                        reader.close();
-                    } catch (final IOException e) {
-                        Log.e(LOG_TAG, "Error closing stream", e);
-                    }
-                }
-            }
-            try {
-                return getWeatherDataFromJson(forecastJsonStr,numDays);
-            } catch (JSONException e) {
-                e.printStackTrace();
-            }
-            return null;
-        }
-    }
 }
